Added Prisma bar list to show fields per model
